Rename tests migrated to use compose to *ComposeTest
