--- datacapture/src/androidTest/java/com/google/android/fhir/datacapture/test/views/DropDownViewHolderFactoryTest.kt
@@ -1,558 +1,0 @@
-/*
- * Copyright 2025 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.android.fhir.datacapture.test.views
-
-import android.widget.FrameLayout
-import android.widget.TextView
-import androidx.compose.ui.semantics.SemanticsProperties
-import androidx.compose.ui.test.SemanticsMatcher
-import androidx.compose.ui.test.assert
-import androidx.compose.ui.test.assertIsDisplayed
-import androidx.compose.ui.test.assertIsNotEnabled
-import androidx.compose.ui.test.assertTextEquals
-import androidx.compose.ui.test.hasAnyDescendant
-import androidx.compose.ui.test.hasTestTag
-import androidx.compose.ui.test.hasTextExactly
-import androidx.compose.ui.test.isPopup
-import androidx.compose.ui.test.junit4.createEmptyComposeRule
-import androidx.compose.ui.test.onNodeWithTag
-import androidx.compose.ui.test.onNodeWithText
-import androidx.compose.ui.test.performClick
-import androidx.compose.ui.text.AnnotatedString
-import androidx.test.ext.junit.rules.ActivityScenarioRule
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.platform.app.InstrumentationRegistry
-import com.google.android.fhir.datacapture.R
-import com.google.android.fhir.datacapture.extensions.displayString
-import com.google.android.fhir.datacapture.extensions.toAnnotatedString
-import com.google.android.fhir.datacapture.test.TestActivity
-import com.google.android.fhir.datacapture.validation.Invalid
-import com.google.android.fhir.datacapture.validation.NotValidated
-import com.google.android.fhir.datacapture.views.QuestionTextConfiguration
-import com.google.android.fhir.datacapture.views.QuestionnaireViewItem
-import com.google.android.fhir.datacapture.views.compose.CLEAR_TEXT_ICON_BUTTON_TAG
-import com.google.android.fhir.datacapture.views.compose.DROP_DOWN_ANSWER_MENU_ITEM_TAG
-import com.google.android.fhir.datacapture.views.compose.DROP_DOWN_TEXT_FIELD_TAG
-import com.google.android.fhir.datacapture.views.compose.ERROR_TEXT_AT_HEADER_TEST_TAG
-import com.google.android.fhir.datacapture.views.factories.DropDownViewHolderFactory
-import com.google.android.fhir.datacapture.views.factories.QuestionnaireItemViewHolder
-import com.google.common.truth.Truth.assertThat
-import org.hl7.fhir.r4.model.Coding
-import org.hl7.fhir.r4.model.Questionnaire
-import org.hl7.fhir.r4.model.QuestionnaireResponse
-import org.hl7.fhir.r4.model.Reference
-import org.junit.Before
-import org.junit.Rule
-import org.junit.Test
-import org.junit.runner.RunWith
-
-@RunWith(AndroidJUnit4::class)
-class DropDownViewHolderFactoryTest {
-  @get:Rule
-  val activityScenarioRule: ActivityScenarioRule<TestActivity> =
-    ActivityScenarioRule(TestActivity::class.java)
-
-  @get:Rule val composeTestRule = createEmptyComposeRule()
-
-  private lateinit var viewHolder: QuestionnaireItemViewHolder
-
-  @Before
-  fun setUp() {
-    activityScenarioRule.scenario.onActivity { activity ->
-      viewHolder = DropDownViewHolderFactory.create(FrameLayout(activity))
-      activity.setContentView(viewHolder.itemView)
-    }
-
-    InstrumentationRegistry.getInstrumentation().waitForIdleSync()
-  }
-
-  @Test
-  fun shouldSetQuestionHeader() {
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent().apply { text = "Question?" },
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-      ),
-    )
-
-    // Synchronize
-    composeTestRule.waitForIdle()
-
-    assertThat(viewHolder.itemView.findViewById<TextView>(R.id.question).text.toString())
-      .isEqualTo("Question?")
-  }
-
-  @Test
-  fun shouldPopulateDropDown() {
-    val answerOption =
-      Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
-        value = Coding().setCode("test-code").setDisplay("Test Code")
-      }
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent().apply { addAnswerOption(answerOption) },
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-      ),
-    )
-    composeTestRule.onNodeWithTag(DROP_DOWN_TEXT_FIELD_TAG).performClick()
-    composeTestRule
-      .onNode(isPopup())
-      .assert(
-        hasAnyDescendant(
-          hasTestTag(DROP_DOWN_ANSWER_MENU_ITEM_TAG) and hasTextExactly("Test Code"),
-        ),
-      )
-  }
-
-  @Test
-  fun shouldPopulateDropdownWithDisplayForReferenceValueType() {
-    val answerOption =
-      Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
-        value =
-          Reference().apply {
-            reference = "Patient/123"
-            display = "John Doe"
-          }
-      }
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent().apply { addAnswerOption(answerOption) },
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-      ),
-    )
-    composeTestRule.onNodeWithTag(DROP_DOWN_TEXT_FIELD_TAG).performClick()
-    composeTestRule
-      .onNode(isPopup())
-      .assert(
-        hasAnyDescendant(
-          hasTestTag(DROP_DOWN_ANSWER_MENU_ITEM_TAG) and hasTextExactly("John Doe"),
-        ),
-      )
-  }
-
-  @Test
-  fun shouldPopulateDropdownWithTypeAndIdForReferenceValueTypeIfMissingDisplay() {
-    val answerOption =
-      Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
-        value = Reference().apply { reference = "Patient/123" }
-      }
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent().apply { addAnswerOption(answerOption) },
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-      ),
-    )
-    composeTestRule.onNodeWithTag(DROP_DOWN_TEXT_FIELD_TAG).performClick()
-    composeTestRule
-      .onNode(isPopup())
-      .assert(
-        hasAnyDescendant(
-          hasTestTag(DROP_DOWN_ANSWER_MENU_ITEM_TAG) and hasTextExactly("Patient/123"),
-        ),
-      )
-  }
-
-  @Test
-  fun shouldSetDropDownOptionToCodeIfValueCodingDisplayEmpty() {
-    val answerOption =
-      Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
-        value = Coding().apply { code = "test-code" }
-      }
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent().apply { addAnswerOption(answerOption) },
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-      ),
-    )
-    composeTestRule.onNodeWithTag(DROP_DOWN_TEXT_FIELD_TAG).performClick()
-    composeTestRule
-      .onNode(isPopup())
-      .assert(
-        hasAnyDescendant(
-          hasTestTag(DROP_DOWN_ANSWER_MENU_ITEM_TAG) and hasTextExactly("test-code"),
-        ),
-      )
-  }
-
-  @Test
-  fun shouldSetSelectedTextEmptyIfAnswerNull() {
-    val answerOption =
-      Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
-        value =
-          Coding().apply {
-            code = "test-code"
-            display = "Test Code"
-          }
-      }
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent().apply { addAnswerOption(answerOption) },
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-      ),
-    )
-
-    composeTestRule
-      .onNodeWithTag(DROP_DOWN_TEXT_FIELD_TAG)
-      .assert(
-        SemanticsMatcher.expectValue(
-          SemanticsProperties.EditableText,
-          AnnotatedString(""),
-        ),
-      )
-  }
-
-  @Test
-  fun shouldSelectedTextToDisplayIfAnswerNotNull() {
-    val answerOption =
-      Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
-        value =
-          Coding().apply {
-            code = "test-code"
-            display = "Test Code"
-          }
-      }
-    val fakeAnswerValueSetResolver = { uri: String ->
-      if (uri == "http://coding-value-set-url") {
-        listOf(answerOption)
-      } else {
-        emptyList()
-      }
-    }
-    val questionnaireItem =
-      Questionnaire.QuestionnaireItemComponent().apply {
-        answerValueSet = "http://coding-value-set-url"
-      }
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        questionnaireItem,
-        QuestionnaireResponse.QuestionnaireResponseItemComponent().apply {
-          addAnswer(
-            QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent().apply {
-              value = answerOption.value
-            },
-          )
-        },
-        enabledAnswerOptions = fakeAnswerValueSetResolver.invoke(questionnaireItem.answerValueSet),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-      ),
-    )
-
-    val context = viewHolder.itemView.context
-    val answerOptionDisplay = answerOption.value.displayString(context)
-    composeTestRule
-      .onNodeWithTag(DROP_DOWN_TEXT_FIELD_TAG)
-      .assert(
-        SemanticsMatcher.expectValue(
-          SemanticsProperties.EditableText,
-          answerOptionDisplay.toAnnotatedString(),
-        ),
-      )
-  }
-
-  @Test
-  fun shouldSelectedTextToDisplayIfAnswerNotNullAndDisplayMatchesMoreThanOneOption() {
-    val answerOption1 =
-      Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
-        value =
-          Reference().apply {
-            reference = "Patient/1234"
-            display = "John"
-          }
-      }
-
-    val answerOption2 =
-      Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
-        value =
-          Reference().apply {
-            reference = "Patient/6789"
-            display = "John"
-          }
-      }
-
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent().apply {
-          addAnswerOption(answerOption1)
-          addAnswerOption(answerOption2)
-        },
-        QuestionnaireResponse.QuestionnaireResponseItemComponent().apply {
-          addAnswer(
-            QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent().apply {
-              value = answerOption2.value
-            },
-          )
-        },
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-      ),
-    )
-
-    val context = viewHolder.itemView.context
-    val answerOption2Display = answerOption2.value.displayString(context)
-    composeTestRule
-      .onNodeWithTag(DROP_DOWN_TEXT_FIELD_TAG)
-      .assert(
-        SemanticsMatcher.expectValue(
-          SemanticsProperties.EditableText,
-          answerOption2Display.toAnnotatedString(),
-        ),
-      )
-  }
-
-  @Test
-  fun displayValidationResult_error_shouldShowErrorMessage() {
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent().apply { required = true },
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
-        validationResult = Invalid(listOf("Missing answer for required field.")),
-        answersChangedCallback = { _, _, _, _ -> },
-      ),
-    )
-    composeTestRule
-      .onNodeWithTag(DROP_DOWN_TEXT_FIELD_TAG)
-      .assert(
-        SemanticsMatcher.expectValue(
-          SemanticsProperties.Error,
-          "Missing answer for required field.",
-        ),
-      )
-  }
-
-  @Test
-  fun displayValidationResult_noError_shouldShowNoErrorMessage() {
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent().apply {
-          required = true
-          addAnswerOption(
-            Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
-              value = Coding().apply { display = "display" }
-            },
-          )
-        },
-        QuestionnaireResponse.QuestionnaireResponseItemComponent()
-          .addAnswer(
-            QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent().apply {
-              value = Coding().apply { display = "display" }
-            },
-          ),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-      ),
-    )
-    composeTestRule
-      .onNodeWithTag(DROP_DOWN_TEXT_FIELD_TAG)
-      .assert(
-        SemanticsMatcher.keyNotDefined(
-          SemanticsProperties.Error,
-        ),
-      )
-  }
-
-  @Test
-  fun hidesErrorTextviewInTheHeader() {
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent(),
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-      ),
-    )
-
-    composeTestRule.onNodeWithTag(ERROR_TEXT_AT_HEADER_TEST_TAG).assertDoesNotExist()
-  }
-
-  @Test
-  fun shouldHideClearIconWhenTextIsEmpty() {
-    val answerOption =
-      Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
-        value =
-          Coding().apply {
-            code = "code"
-            display = "display"
-          }
-      }
-
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent().apply { addAnswerOption(answerOption) },
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-      ),
-    )
-    composeTestRule.onNodeWithTag(CLEAR_TEXT_ICON_BUTTON_TAG).assertDoesNotExist()
-  }
-
-  @Test
-  fun shouldShowClearIconWhenTextIsNotEmpty() {
-    val answerOption =
-      Questionnaire.QuestionnaireItemAnswerOptionComponent().apply {
-        value =
-          Coding().apply {
-            code = "code"
-            display = "display"
-          }
-      }
-
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent().apply { addAnswerOption(answerOption) },
-        QuestionnaireResponse.QuestionnaireResponseItemComponent().apply {
-          addAnswer().apply { value = answerOption.valueCoding }
-        },
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-      ),
-    )
-
-    composeTestRule.onNodeWithTag(CLEAR_TEXT_ICON_BUTTON_TAG).assertIsDisplayed()
-  }
-
-  @Test
-  fun bind_readOnly_shouldDisableView() {
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent().apply { readOnly = true },
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-      ),
-    )
-    composeTestRule.onNodeWithTag(DROP_DOWN_TEXT_FIELD_TAG).assertIsNotEnabled()
-  }
-
-  @Test
-  fun showsAsterisk() {
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent().apply {
-          text = "Question?"
-          required = true
-        },
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-        questionViewTextConfiguration = QuestionTextConfiguration(showAsterisk = true),
-      ),
-    )
-    // Synchronize
-    composeTestRule.waitForIdle()
-
-    assertThat(viewHolder.itemView.findViewById<TextView>(R.id.question).text.toString())
-      .isEqualTo("Question? *")
-  }
-
-  @Test
-  fun hideAsterisk() {
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent().apply {
-          text = "Question?"
-          required = true
-        },
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-        questionViewTextConfiguration = QuestionTextConfiguration(showAsterisk = false),
-      ),
-    )
-    // Synchronize
-    composeTestRule.waitForIdle()
-    assertThat(viewHolder.itemView.findViewById<TextView>(R.id.question).text.toString())
-      .isEqualTo("Question?")
-  }
-
-  @Test
-  fun showsRequiredText() {
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent().apply { required = true },
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-        questionViewTextConfiguration = QuestionTextConfiguration(showRequiredText = true),
-      ),
-    )
-
-    composeTestRule
-      .onNodeWithTag(DROP_DOWN_TEXT_FIELD_TAG)
-      .assertTextEquals("Required", includeEditableText = false)
-  }
-
-  @Test
-  fun hideRequiredText() {
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent().apply { required = true },
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-        questionViewTextConfiguration = QuestionTextConfiguration(showRequiredText = false),
-      ),
-    )
-
-    composeTestRule.onNodeWithText("Required", substring = true).assertDoesNotExist()
-  }
-
-  @Test
-  fun showsOptionalText() {
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent(),
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-        questionViewTextConfiguration = QuestionTextConfiguration(showOptionalText = true),
-      ),
-    )
-
-    composeTestRule
-      .onNodeWithTag(DROP_DOWN_TEXT_FIELD_TAG)
-      .assertTextEquals("Optional", includeEditableText = false)
-  }
-
-  @Test
-  fun hideOptionalText() {
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent(),
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-        questionViewTextConfiguration = QuestionTextConfiguration(showOptionalText = false),
-      ),
-    )
-
-    composeTestRule.onNodeWithText("Optional", substring = true).assertDoesNotExist()
-  }
-}
