+++ datacapture/src/androidTest/java/com/google/android/fhir/datacapture/test/views/EditTextIntegerViewHolderFactoryComposeTest.kt
@@ -52,7 +52,7 @@ import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class EditTextIntegerViewHolderFactoryInstrumentedTest {
+class EditTextIntegerViewHolderFactoryComposeTest {
 
   @get:Rule
   val activityScenarioRule: ActivityScenarioRule<TestActivity> =
