+++ datacapture/src/androidTest/java/com/google/android/fhir/datacapture/test/views/EditTextSingleLineViewHolderFactoryComposeTest.kt
@@ -53,7 +53,7 @@ import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class EditTextSingleLineViewHolderFactoryInstrumentedTest {
+class EditTextSingleLineViewHolderFactoryComposeTest {
 
   @get:Rule
   val activityScenarioRule: ActivityScenarioRule<TestActivity> =
