+++ datacapture/src/androidTest/java/com/google/android/fhir/datacapture/test/views/EditTextViewHolderDelegateComposeTest.kt
@@ -18,8 +18,6 @@ package com.google.android.fhir.datacapture.test.views
 
 import android.widget.FrameLayout
 import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.assertTextEquals
 import androidx.compose.ui.test.junit4.createEmptyComposeRule
@@ -49,7 +47,7 @@ import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class EditTextViewHolderDelegateTest {
+class EditTextViewHolderDelegateComposeTest {
 
   @get:Rule
   val activityScenarioRule: ActivityScenarioRule<TestActivity> =
