+++ datacapture/src/androidTest/java/com/google/android/fhir/datacapture/test/views/QuantityViewHolderFactoryComposeTest.kt
@@ -18,16 +18,19 @@ package com.google.android.fhir.datacapture.test.views
 
 import android.widget.FrameLayout
 import android.widget.TextView
-import androidx.compose.ui.test.IdlingResource
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.assertIsNotDisplayed
 import androidx.compose.ui.test.assertIsNotEnabled
 import androidx.compose.ui.test.assertTextEquals
+import androidx.compose.ui.test.hasAnyAncestor
+import androidx.compose.ui.test.hasText
+import androidx.compose.ui.test.isPopup
 import androidx.compose.ui.test.junit4.createEmptyComposeRule
 import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.onNodeWithText
-import androidx.compose.ui.test.performTextReplacement
+import androidx.compose.ui.test.performClick
+import androidx.compose.ui.test.performTextInput
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
@@ -35,28 +38,28 @@ import com.google.android.fhir.datacapture.R
 import com.google.android.fhir.datacapture.test.TestActivity
 import com.google.android.fhir.datacapture.validation.Invalid
 import com.google.android.fhir.datacapture.validation.NotValidated
+import com.google.android.fhir.datacapture.validation.Valid
 import com.google.android.fhir.datacapture.views.QuestionTextConfiguration
 import com.google.android.fhir.datacapture.views.QuestionnaireViewItem
+import com.google.android.fhir.datacapture.views.compose.DROP_DOWN_TEXT_FIELD_TAG
 import com.google.android.fhir.datacapture.views.compose.EDIT_TEXT_FIELD_TEST_TAG
 import com.google.android.fhir.datacapture.views.compose.ERROR_TEXT_AT_HEADER_TEST_TAG
-import com.google.android.fhir.datacapture.views.compose.UNIT_TEXT_TEST_TAG
-import com.google.android.fhir.datacapture.views.factories.EditTextDecimalViewHolderFactory
+import com.google.android.fhir.datacapture.views.factories.QuantityViewHolderFactory
 import com.google.android.fhir.datacapture.views.factories.QuestionnaireItemViewHolder
 import com.google.common.truth.Truth.assertThat
 import java.math.BigDecimal
 import org.hl7.fhir.r4.model.Coding
-import org.hl7.fhir.r4.model.DecimalType
 import org.hl7.fhir.r4.model.Extension
+import org.hl7.fhir.r4.model.Quantity
 import org.hl7.fhir.r4.model.Questionnaire
 import org.hl7.fhir.r4.model.QuestionnaireResponse
-import org.junit.After
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class EditTextDecimalViewHolderFactoryInstrumentedTest {
+class QuantityViewHolderFactoryComposeTest {
 
   @get:Rule
   val activityScenarioRule: ActivityScenarioRule<TestActivity> =
@@ -65,34 +68,19 @@ class EditTextDecimalViewHolderFactoryInstrumentedTest {
   @get:Rule val composeTestRule = createEmptyComposeRule()
 
   private lateinit var viewHolder: QuestionnaireItemViewHolder
-  private lateinit var parent: FrameLayout
-
-  private var pendingTextChange = 0
-  private val handlingTextIdlingResource =
-    object : IdlingResource {
-      override val isIdleNow: Boolean
-        get() = pendingTextChange == 0
-    }
 
   @Before
-  fun setup() {
+  fun setUp() {
     activityScenarioRule.scenario.onActivity { activity ->
-      parent = FrameLayout(activity)
-      viewHolder = EditTextDecimalViewHolderFactory.create(parent)
+      viewHolder = QuantityViewHolderFactory.create(FrameLayout(activity))
       activity.setContentView(viewHolder.itemView)
     }
+
     InstrumentationRegistry.getInstrumentation().waitForIdleSync()
-
-    composeTestRule.registerIdlingResource(handlingTextIdlingResource)
-  }
-
-  @After
-  fun tearDown() {
-    composeTestRule.unregisterIdlingResource(handlingTextIdlingResource)
   }
 
   @Test
-  fun shouldSetQuestionnaireHeader() {
+  fun shouldSetQuestionText() {
     viewHolder.bind(
       QuestionnaireViewItem(
         Questionnaire.QuestionnaireItemComponent().apply { text = "Question?" },
@@ -101,6 +89,7 @@ class EditTextDecimalViewHolderFactoryInstrumentedTest {
         answersChangedCallback = { _, _, _, _ -> },
       ),
     )
+
     // Synchronize
     composeTestRule.waitForIdle()
 
@@ -109,14 +98,14 @@ class EditTextDecimalViewHolderFactoryInstrumentedTest {
   }
 
   @Test
-  fun shouldSetInputText() {
+  fun shouldSetInputDecimalValue() {
     viewHolder.bind(
       QuestionnaireViewItem(
         Questionnaire.QuestionnaireItemComponent(),
         QuestionnaireResponse.QuestionnaireResponseItemComponent().apply {
           addAnswer(
             QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent().apply {
-              value = DecimalType("1.1")
+              value = Quantity().apply { value = BigDecimal("5") }
             },
           )
         },
@@ -124,18 +113,19 @@ class EditTextDecimalViewHolderFactoryInstrumentedTest {
         answersChangedCallback = { _, _, _, _ -> },
       ),
     )
-    composeTestRule.onNodeWithTag(EDIT_TEXT_FIELD_TEST_TAG).assertTextEquals("1.1")
+
+    composeTestRule.onNodeWithTag(EDIT_TEXT_FIELD_TEST_TAG).assertTextEquals("5")
   }
 
   @Test
-  fun shouldSetInputTextToEmpty() {
+  fun shouldClearInputDecimalValue() {
     viewHolder.bind(
       QuestionnaireViewItem(
         Questionnaire.QuestionnaireItemComponent(),
         QuestionnaireResponse.QuestionnaireResponseItemComponent().apply {
           addAnswer(
             QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent().apply {
-              value = DecimalType("1.1")
+              value = Quantity().apply { value = BigDecimal("5") }
             },
           )
         },
@@ -143,6 +133,7 @@ class EditTextDecimalViewHolderFactoryInstrumentedTest {
         answersChangedCallback = { _, _, _, _ -> },
       ),
     )
+    composeTestRule.onNodeWithTag(EDIT_TEXT_FIELD_TEST_TAG).assertTextEquals("5")
 
     viewHolder.bind(
       QuestionnaireViewItem(
@@ -152,27 +143,18 @@ class EditTextDecimalViewHolderFactoryInstrumentedTest {
         answersChangedCallback = { _, _, _, _ -> },
       ),
     )
-
     composeTestRule.onNodeWithTag(EDIT_TEXT_FIELD_TEST_TAG).assertTextEquals("")
   }
 
   @Test
-  fun shouldSetUnitText() {
+  fun shouldSetUnitValue() {
     viewHolder.bind(
       QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent().apply {
-          type = Questionnaire.QuestionnaireItemType.DECIMAL
-          addExtension(
-            Extension().apply {
-              url = "http://hl7.org/fhir/StructureDefinition/questionnaire-unit"
-              setValue(Coding().apply { code = "kg" })
-            },
-          )
-        },
+        Questionnaire.QuestionnaireItemComponent(),
         QuestionnaireResponse.QuestionnaireResponseItemComponent().apply {
           addAnswer(
             QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent().apply {
-              value = DecimalType("1.1")
+              value = Quantity().apply { unit = "kg" }
             },
           )
         },
@@ -181,26 +163,41 @@ class EditTextDecimalViewHolderFactoryInstrumentedTest {
       ),
     )
 
-    composeTestRule.onNodeWithTag(UNIT_TEXT_TEST_TAG).assertTextEquals("kg")
+    composeTestRule.onNodeWithTag(DROP_DOWN_TEXT_FIELD_TAG).assertTextEquals("kg")
   }
 
   @Test
-  fun shouldClearUnitText() {
+  fun shouldSetUnitValueFromInitialWhenAnswerIsMissing() {
     viewHolder.bind(
       QuestionnaireViewItem(
         Questionnaire.QuestionnaireItemComponent().apply {
-          type = Questionnaire.QuestionnaireItemType.DECIMAL
-          addExtension(
-            Extension().apply {
-              url = "http://hl7.org/fhir/StructureDefinition/questionnaire-unit"
-              setValue(Coding().apply { code = "kg" })
-            },
+          addInitial(
+            Questionnaire.QuestionnaireItemInitialComponent(
+              Quantity().apply {
+                this.unit = "kg"
+                this.code = "kilo"
+              },
+            ),
           )
         },
+        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
+        validationResult = NotValidated,
+        answersChangedCallback = { _, _, _, _ -> },
+      ),
+    )
+
+    composeTestRule.onNodeWithTag(DROP_DOWN_TEXT_FIELD_TAG).assertTextEquals("kg")
+  }
+
+  @Test
+  fun shouldClearUnitValue() {
+    viewHolder.bind(
+      QuestionnaireViewItem(
+        Questionnaire.QuestionnaireItemComponent(),
         QuestionnaireResponse.QuestionnaireResponseItemComponent().apply {
           addAnswer(
             QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent().apply {
-              value = DecimalType("1.1")
+              value = Quantity().apply { unit = "kg" }
             },
           )
         },
@@ -208,6 +205,9 @@ class EditTextDecimalViewHolderFactoryInstrumentedTest {
         answersChangedCallback = { _, _, _, _ -> },
       ),
     )
+
+    composeTestRule.onNodeWithTag(DROP_DOWN_TEXT_FIELD_TAG).assertTextEquals("kg")
+
     viewHolder.bind(
       QuestionnaireViewItem(
         Questionnaire.QuestionnaireItemComponent(),
@@ -217,91 +217,37 @@ class EditTextDecimalViewHolderFactoryInstrumentedTest {
       ),
     )
 
-    composeTestRule.onNodeWithTag(EDIT_TEXT_FIELD_TEST_TAG).assertTextEquals("")
+    composeTestRule.onNodeWithTag(DROP_DOWN_TEXT_FIELD_TAG).assertTextEquals("")
   }
 
   @Test
-  fun shouldSetQuestionnaireResponseItemAnswerIfTextIsValid() {
-    var answers: List<QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent>? = null
-    val questionnaireViewItem =
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent(),
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, result, _ ->
-          answers = result
-          pendingTextChange -= if (pendingTextChange > 0) 1 else 0
-        },
-      )
-
-    viewHolder.bind(questionnaireViewItem)
-    composeTestRule.onNodeWithTag(EDIT_TEXT_FIELD_TEST_TAG).performTextReplacement("1.1").also {
-      pendingTextChange += 1
-    }
-    composeTestRule.waitForIdle()
-
-    assertThat(answers!!.single().valueDecimalType.value).isEqualTo(BigDecimal.valueOf(1.1))
-  }
-
-  @Test
-  fun shouldSetQuestionnaireResponseItemAnswerToEmpty() {
-    val questionnaireViewItem =
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent(),
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-      )
-    viewHolder.bind(questionnaireViewItem)
-    composeTestRule.onNodeWithTag(EDIT_TEXT_FIELD_TEST_TAG).performTextReplacement("")
-    composeTestRule.waitForIdle()
-
-    assertThat(questionnaireViewItem.answers).isEmpty()
-  }
-
-  @Test
-  fun shouldSetDraftAnswerIfTextIsInvalid() {
-    var draftAnswer: Any? = null
-    val questionnaireViewItem =
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent(),
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, result ->
-          draftAnswer = result
-          pendingTextChange -= if (pendingTextChange > 0) 1 else 0
-        },
-      )
-    viewHolder.bind(questionnaireViewItem)
-    composeTestRule.onNodeWithTag(EDIT_TEXT_FIELD_TEST_TAG).performTextReplacement("1.1.1.1").also {
-      pendingTextChange += 1
-    }
-    composeTestRule.waitForIdle()
-    assertThat(draftAnswer as String).isEqualTo("1.1.1.1")
-  }
-
-  @Test
-  fun displayValidationResultShouldShowNoErrorMesssage() {
+  fun shouldDisplayErrorMessageInValidationResult() {
     viewHolder.bind(
       QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent().apply {
-          addExtension().apply {
-            url = "http://hl7.org/fhir/StructureDefinition/minValue"
-            setValue(DecimalType("2.2"))
-          }
-          addExtension().apply {
-            url = "http://hl7.org/fhir/StructureDefinition/maxValue"
-            setValue(DecimalType("4.4"))
-          }
-        },
+        Questionnaire.QuestionnaireItemComponent().apply { required = true },
+        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
+        validationResult = Invalid(listOf("Missing answer for required field.")),
+        answersChangedCallback = { _, _, _, _ -> },
+      ),
+    )
+
+    composeTestRule.onNodeWithContentDescription("Error").assertIsDisplayed()
+    composeTestRule.onNodeWithText("Missing answer for required field.").assertIsDisplayed()
+  }
+
+  @Test
+  fun shouldDisplayNoErrorMessageWhenValidationResultIsValid() {
+    viewHolder.bind(
+      QuestionnaireViewItem(
+        Questionnaire.QuestionnaireItemComponent().apply { required = true },
         QuestionnaireResponse.QuestionnaireResponseItemComponent().apply {
           addAnswer(
             QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent().apply {
-              value = DecimalType("3.3")
+              value = Quantity(22.5)
             },
           )
         },
-        validationResult = NotValidated,
+        validationResult = Valid,
         answersChangedCallback = { _, _, _, _ -> },
       ),
     )
@@ -310,51 +256,7 @@ class EditTextDecimalViewHolderFactoryInstrumentedTest {
   }
 
   @Test
-  fun displayValidationResultShouldShowErrorMessage() {
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent().apply {
-          addExtension().apply {
-            url = "http://hl7.org/fhir/StructureDefinition/minValue"
-            setValue(DecimalType("2.1"))
-          }
-          addExtension().apply {
-            url = "http://hl7.org/fhir/StructureDefinition/maxValue"
-            setValue(DecimalType("4.2"))
-          }
-        },
-        QuestionnaireResponse.QuestionnaireResponseItemComponent().apply {
-          addAnswer(
-            QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent().apply {
-              value = DecimalType("1.1")
-            },
-          )
-        },
-        validationResult = Invalid(listOf("Minimum value allowed is:2.1")),
-        answersChangedCallback = { _, _, _, _ -> },
-      ),
-    )
-
-    composeTestRule.onNodeWithContentDescription("Error").assertIsDisplayed()
-    composeTestRule.onNodeWithText("Minimum value allowed is:2.1").assertIsDisplayed()
-  }
-
-  @Test
-  fun hidesErrorTextviewInTheHeader() {
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent(),
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-      ),
-    )
-
-    composeTestRule.onNodeWithTag(ERROR_TEXT_AT_HEADER_TEST_TAG).assertIsNotDisplayed()
-  }
-
-  @Test
-  fun bindReadOnlyShouldDisableView() {
+  fun shouldDisableTextInputInReadOnlyMode() {
     viewHolder.bind(
       QuestionnaireViewItem(
         Questionnaire.QuestionnaireItemComponent().apply { readOnly = true },
@@ -364,11 +266,151 @@ class EditTextDecimalViewHolderFactoryInstrumentedTest {
       ),
     )
 
-    composeTestRule.onNodeWithTag(EDIT_TEXT_FIELD_TEST_TAG).assertIsDisplayed().assertIsNotEnabled()
+    composeTestRule.onNodeWithTag(EDIT_TEXT_FIELD_TEST_TAG).assertIsNotEnabled()
   }
 
   @Test
-  fun showAsterisk() {
+  fun shouldDisableUnitInputInReadOnlyMode() {
+    viewHolder.bind(
+      QuestionnaireViewItem(
+        Questionnaire.QuestionnaireItemComponent().apply { readOnly = true },
+        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
+        validationResult = NotValidated,
+        answersChangedCallback = { _, _, _, _ -> },
+      ),
+    )
+
+    composeTestRule.onNodeWithTag(DROP_DOWN_TEXT_FIELD_TAG).assertIsNotEnabled()
+  }
+
+  @Test
+  fun shouldAlwaysHideErrorTextviewInTheHeader() {
+    viewHolder.bind(
+      QuestionnaireViewItem(
+        Questionnaire.QuestionnaireItemComponent(),
+        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
+        validationResult = NotValidated,
+        answersChangedCallback = { _, _, _, _ -> },
+      ),
+    )
+
+    composeTestRule
+      .onNodeWithTag(ERROR_TEXT_AT_HEADER_TEST_TAG)
+      .assertIsNotDisplayed()
+      .assertDoesNotExist()
+  }
+
+  @Test
+  fun shouldSetDraftWithUnit() {
+    var answerHolder: List<QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent>? = null
+    var draftHolder: Any? = null
+
+    val questionnaireViewItem = createQuestionnaireViewItem { answers, draft ->
+      answerHolder = answers
+      draftHolder = draft
+    }
+
+    viewHolder.bind(questionnaireViewItem)
+    composeTestRule.onNodeWithTag(DROP_DOWN_TEXT_FIELD_TAG).performClick()
+    composeTestRule
+      .onNode(hasText("centimeter") and hasAnyAncestor(isPopup()))
+      .assertIsDisplayed()
+      .performClick()
+    composeTestRule.onNodeWithTag(DROP_DOWN_TEXT_FIELD_TAG).assertTextEquals("centimeter")
+
+    composeTestRule.waitUntil { draftHolder != null }
+
+    with(draftHolder as Coding) {
+      assertThat(system).isEqualTo("http://unitofmeasure.com")
+      assertThat(code).isEqualTo("cm")
+      assertThat(display).isEqualTo("centimeter")
+    }
+    assertThat(answerHolder).isEmpty()
+  }
+
+  @Test
+  fun shouldSetDraftWithDecimalValue() {
+    var answerHolder: List<QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent>? = null
+    var draftHolder: Any? = null
+
+    val questionnaireViewItem = createQuestionnaireViewItem { answers, draft ->
+      answerHolder = answers
+      draftHolder = draft
+    }
+
+    viewHolder.bind(questionnaireViewItem)
+
+    composeTestRule.onNodeWithTag(EDIT_TEXT_FIELD_TEST_TAG).performClick().performTextInput("22")
+    composeTestRule.waitUntil { draftHolder != null }
+    composeTestRule.onNodeWithTag(EDIT_TEXT_FIELD_TEST_TAG).assertTextEquals("22")
+
+    assertThat(draftHolder as BigDecimal).isEqualTo(BigDecimal(22))
+    assertThat(answerHolder).isEmpty()
+  }
+
+  @Test
+  fun draftWithUnit_shouldCompleteQuantity() {
+    var answerHolder: List<QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent>? = null
+    var draftHolder: Any? = null
+
+    val questionnaireViewItem =
+      createQuestionnaireViewItem(Coding("http://unitofmeasure.com", "cm", "centimeter")) {
+        answers,
+        draft,
+        ->
+        answerHolder = answers
+        draftHolder = draft
+      }
+
+    viewHolder.bind(questionnaireViewItem)
+    composeTestRule.onNodeWithTag(DROP_DOWN_TEXT_FIELD_TAG).performClick()
+    composeTestRule.onNodeWithTag(EDIT_TEXT_FIELD_TEST_TAG).performClick().performTextInput("22")
+
+    composeTestRule.waitUntil { !answerHolder.isNullOrEmpty() }
+
+    composeTestRule.onNodeWithTag(EDIT_TEXT_FIELD_TEST_TAG).assertTextEquals("22")
+
+    with(answerHolder!!.single().valueQuantity) {
+      assertThat(system).isEqualTo("http://unitofmeasure.com")
+      assertThat(code).isEqualTo("cm")
+      assertThat(unit).isEqualTo("centimeter")
+      assertThat(value).isEqualTo(BigDecimal("22.0"))
+    }
+    assertThat(draftHolder).isNull()
+  }
+
+  @Test
+  fun draftWithDecimalValue_shouldCompleteQuantity() {
+    var answerHolder: List<QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent>? = null
+    var draftHolder: Any? = null
+
+    val questionnaireViewItem =
+      createQuestionnaireViewItem(BigDecimal(22)) { answers, draft ->
+        answerHolder = answers
+        draftHolder = draft
+      }
+
+    viewHolder.bind(questionnaireViewItem)
+    composeTestRule.onNodeWithTag(DROP_DOWN_TEXT_FIELD_TAG).performClick()
+    composeTestRule
+      .onNode(hasText("centimeter") and hasAnyAncestor(isPopup()))
+      .assertIsDisplayed()
+      .performClick()
+    composeTestRule.onNodeWithTag(DROP_DOWN_TEXT_FIELD_TAG).assertTextEquals("centimeter")
+
+    composeTestRule.waitUntil { !answerHolder.isNullOrEmpty() }
+
+    with(answerHolder!!.single().valueQuantity) {
+      assertThat(system).isEqualTo("http://unitofmeasure.com")
+      assertThat(code).isEqualTo("cm")
+      assertThat(unit).isEqualTo("centimeter")
+      assertThat(value).isEqualTo(BigDecimal("22.0"))
+    }
+    assertThat(draftHolder).isNull()
+  }
+
+  @Test
+  fun shouldShowAsterisk() {
     viewHolder.bind(
       QuestionnaireViewItem(
         Questionnaire.QuestionnaireItemComponent().apply {
@@ -381,7 +423,6 @@ class EditTextDecimalViewHolderFactoryInstrumentedTest {
         questionViewTextConfiguration = QuestionTextConfiguration(showAsterisk = true),
       ),
     )
-
     // Synchronize
     composeTestRule.waitForIdle()
 
@@ -390,7 +431,7 @@ class EditTextDecimalViewHolderFactoryInstrumentedTest {
   }
 
   @Test
-  fun hideAsterisk() {
+  fun shouldHideAsterisk() {
     viewHolder.bind(
       QuestionnaireViewItem(
         Questionnaire.QuestionnaireItemComponent().apply {
@@ -412,7 +453,7 @@ class EditTextDecimalViewHolderFactoryInstrumentedTest {
   }
 
   @Test
-  fun showsRequiredText() {
+  fun shouldShowRequiredText() {
     viewHolder.bind(
       QuestionnaireViewItem(
         Questionnaire.QuestionnaireItemComponent().apply { required = true },
@@ -427,7 +468,7 @@ class EditTextDecimalViewHolderFactoryInstrumentedTest {
   }
 
   @Test
-  fun hideRequiredText() {
+  fun shouldHideRequiredText() {
     viewHolder.bind(
       QuestionnaireViewItem(
         Questionnaire.QuestionnaireItemComponent().apply { required = true },
@@ -437,11 +478,12 @@ class EditTextDecimalViewHolderFactoryInstrumentedTest {
         questionViewTextConfiguration = QuestionTextConfiguration(showRequiredText = false),
       ),
     )
+
     composeTestRule.onNodeWithText("Required").assertDoesNotExist()
   }
 
   @Test
-  fun showOptionalText() {
+  fun shouldShowOptionalText() {
     viewHolder.bind(
       QuestionnaireViewItem(
         Questionnaire.QuestionnaireItemComponent(),
@@ -456,7 +498,7 @@ class EditTextDecimalViewHolderFactoryInstrumentedTest {
   }
 
   @Test
-  fun hideOptionalText() {
+  fun shouldHideOptionalText() {
     viewHolder.bind(
       QuestionnaireViewItem(
         Questionnaire.QuestionnaireItemComponent(),
@@ -470,32 +512,43 @@ class EditTextDecimalViewHolderFactoryInstrumentedTest {
     composeTestRule.onNodeWithText("Optional").assertDoesNotExist()
   }
 
-  @Test
-  fun bindAgainShouldRemovePreviousText() {
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent(),
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-        draftAnswer = "1.1.1.1",
-      ),
+  private fun createQuestionnaireViewItem(
+    draftAnswer: Any? = null,
+    answersChangedCallback:
+      (List<QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent>, Any?) -> Unit,
+  ): QuestionnaireViewItem {
+    return QuestionnaireViewItem(
+      Questionnaire.QuestionnaireItemComponent().apply {
+        required = true
+        addExtension(
+          Extension().apply {
+            url = "http://hl7.org/fhir/StructureDefinition/questionnaire-unitOption"
+            setValue(
+              Coding().apply {
+                code = "cm"
+                system = "http://unitofmeasure.com"
+                display = "centimeter"
+              },
+            )
+          },
+        )
+        addExtension(
+          Extension().apply {
+            url = "http://hl7.org/fhir/StructureDefinition/questionnaire-unitOption"
+            setValue(
+              Coding().apply {
+                code = "[in_i]"
+                system = "http://unitofmeasure.com"
+                display = "inch"
+              },
+            )
+          },
+        )
+      },
+      QuestionnaireResponse.QuestionnaireResponseItemComponent(),
+      validationResult = NotValidated,
+      answersChangedCallback = { _, _, answers, draft -> answersChangedCallback(answers, draft) },
+      draftAnswer = draftAnswer,
     )
-
-    composeTestRule
-      .onNodeWithTag(EDIT_TEXT_FIELD_TEST_TAG, useUnmergedTree = true)
-      .assertTextEquals("1.1.1.1")
-    viewHolder.bind(
-      QuestionnaireViewItem(
-        Questionnaire.QuestionnaireItemComponent(),
-        QuestionnaireResponse.QuestionnaireResponseItemComponent(),
-        validationResult = NotValidated,
-        answersChangedCallback = { _, _, _, _ -> },
-      ),
-    )
-
-    composeTestRule
-      .onNodeWithTag(EDIT_TEXT_FIELD_TEST_TAG, useUnmergedTree = true)
-      .assertTextEquals("")
   }
 }
