+++ datacapture/src/androidTest/java/com/google/android/fhir/datacapture/test/views/compose/DropDownItemComposeTest.kt
@@ -35,7 +35,7 @@ import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class DropDownItemTest {
+class DropDownItemComposeTest {
 
   @get:Rule val composeTestRule = createComposeRule()
