+++ datacapture/src/androidTest/java/com/google/android/fhir/datacapture/test/views/compose/HeaderComposeTest.kt
@@ -57,7 +57,7 @@ import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class HeaderTest {
+class HeaderComposeTest {
 
   @get:Rule val composeTestRule = createAndroidComposeRule<ComponentActivity>()
