+++ datacapture/src/androidTest/java/com/google/android/fhir/datacapture/test/views/compose/MediaItemComposeTest.kt
@@ -30,7 +30,7 @@ import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class MediaItemTest {
+class MediaItemComposeTest {
 
   @get:Rule val composeTestRule = createAndroidComposeRule<ComponentActivity>()
